вынес все, кроме канала в абстрактный FixClientBase, чтобы была возможность иначе обрабатывать сообщения, кроме как через канал
