+++ Version.props
@@ -1,5 +1,5 @@
 <Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
     <PropertyGroup>
-        <Version>7.4.0.3</Version>
+        <Version>7.5.0.0</Version>
     </PropertyGroup>
 </Project>
