+++ src/Fix.Initiator/DependencyInjection/ServiceCollectionExtensions.cs
@@ -1,4 +1,3 @@
-using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using QuickFix;
