+++ src/Fix.Initiator/IFixApplication.cs
@@ -1,0 +1,8 @@
+using QuickFix;
+
+namespace SoftWell.Fix.Initiator;
+
+public interface IFixApplication : IApplication
+{
+    SessionSettings SessionSettings { get; }
+}
