+++ src/Fix.Initiator/IFixClient.cs
@@ -1,8 +1,3 @@
-using QuickFix;
-
 namespace SoftWell.Fix.Initiator;
 
-public interface IFixClient : IApplication, IFixMessagesReader, IFixMessagesSender
-{
-    SessionSettings SessionSettings { get; }
-}
+public interface IFixClient : IFixApplication, IFixMessagesReader, IFixMessagesSender { }
